+++ src/utils/validator.ts
@@ -1,4 +1,14 @@
+export const validateName = (name: string): boolean => {
+  return name.length >= 2;
+}
+
 export const validateEmail = (email: string): boolean => {
   const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
   return emailRegex.test(email);
 };
+
+// 비밀번호 유효성 검사 (숫자, 영문, 특수문자 포함 8 ~ 16자)
+export const validatePassword = (password: string): boolean => {
+  const passwordRegex = /^(?=.*[a-zA-Z])(?=.*[0-9])(?=.*[!@#$%^&*])[A-Za-z\d!@#$%^&*]{8,16}$/;
+  return passwordRegex.test(password);
+}
